Fix: savegames work again
The number of unit and building infos are now both saved and loaded as ints.
Ints always use 4 bytes while the enum types will use 1 or 2 bytes (2 in this case)

+++ Project Files/DLLSources/CvInitCore.cpp
@@ -2130,8 +2130,8 @@ void CvInitCore::write(FDataStreamBase* pStream)
 
 	m_aeSlotStatus[getActivePlayer()] = eOldStatus;
 
-	writer.Write(NUM_UNIT_TYPES);
-	writer.Write(NUM_BUILDING_TYPES);
+	writer.Write((int)NUM_UNIT_TYPES);
+	writer.Write((int)NUM_BUILDING_TYPES);
 
 	writerbase.WriteFile();
 }
